+++ package.json
@@ -1,30 +1,31 @@
 {
   "urls": [
-    ["lib/fifo.py", "http://localhost:8000/pico-lib/fifo.py"],
-    ["lib/piotimer.py", "http://localhost:8000/pico-lib/piotimer.py"],
-    ["lib/ssd1306.mpy", "http://localhost:8000/pico-lib/ssd1306.mpy"],
-    ["lib/umqtt/simple.mpy", "http://localhost:8000/pico-lib/umqtt/simple.mpy"],
+    ["lib/fifo.py", "http://localhost:8000/src/pico-lib/fifo.py"],
+    ["lib/piotimer.py", "http://localhost:8000/src/pico-lib/piotimer.py"],
+    ["lib/ssd1306.mpy", "http://localhost:8000/src/pico-lib/ssd1306.mpy"],
+    ["lib/umqtt/simple.mpy", "http://localhost:8000/src/pico-lib/umqtt/simple.mpy"],
 
-    ["data_processing.py", "http://localhost:8000/data_processing.py"],
-    ["data_structure.py", "http://localhost:8000/data_structure.py"],
-    ["hardware.py", "http://localhost:8000/hardware.py"],
     ["main.py", "http://localhost:8000/main.py"],
-    ["main_menu.py", "http://localhost:8000/main_menu.py"],
-    ["measure.py", "http://localhost:8000/measure.py"],
-    ["measure_analysis.py", "http://localhost:8000/measure_analysis.py"],
-    ["pico_network.py", "http://localhost:8000/pico_network.py"],
-    ["result.py", "http://localhost:8000/result.py"],
-    ["save_system.py", "http://localhost:8000/save_system.py"],
-    ["settings.py", "http://localhost:8000/settings.py"],
-    ["state.py", "http://localhost:8000/state.py"],
-    ["state_machine.py", "http://localhost:8000/state_machine.py"],
-    ["utils.py", "http://localhost:8000/utils.py"],
-    ["view.py", "http://localhost:8000/view.py"],
     ["config.json", "http://localhost:8000/config.json"],
 
-    ["resources/animation_power_on.py", "http://localhost:8000/resources/animation_power_on.py"],
-    ["resources/pic_icon.py", "http://localhost:8000/resources/pic_icon.py"],
-    ["resources/pic_loading_circle.py", "http://localhost:8000/resources/pic_loading_circle.py"]
+    ["src/data_processing.py", "http://localhost:8000/src/data_processing.py"],
+    ["src/data_structure.py", "http://localhost:8000/src/data_structure.py"],
+    ["src/hardware.py", "http://localhost:8000/src/hardware.py"],
+    ["src/main_menu.py", "http://localhost:8000/src/main_menu.py"],
+    ["src/measure.py", "http://localhost:8000/src/measure.py"],
+    ["src/measure_analysis.py", "http://localhost:8000/src/measure_analysis.py"],
+    ["src/pico_network.py", "http://localhost:8000/src/pico_network.py"],
+    ["src/result.py", "http://localhost:8000/src/result.py"],
+    ["src/save_system.py", "http://localhost:8000/src/save_system.py"],
+    ["src/settings.py", "http://localhost:8000/src/settings.py"],
+    ["src/state.py", "http://localhost:8000/src/state.py"],
+    ["src/state_machine.py", "http://localhost:8000/src/state_machine.py"],
+    ["src/utils.py", "http://localhost:8000/src/utils.py"],
+    ["src/view.py", "http://localhost:8000/src/view.py"],
+
+    ["src/res/animation_power_on.py", "http://localhost:8000/src/res/animation_power_on.py"],
+    ["src/res/pic_icon.py", "http://localhost:8000/src/res/pic_icon.py"],
+    ["src/res/pic_loading_circle.py", "http://localhost:8000/src/res/pic_loading_circle.py"]
   ],
   "deps": [
   ],
