+++ src/pico_network.py
@@ -1,5 +1,5 @@
 import network
-from utils import GlobalSettings, print_log
+from src.utils import GlobalSettings, print_log
 import json
 from umqtt.simple import MQTTClient
