+++ src/res/pic_icon.py
