+++ src/res/pic_loading_circle.py
