+++ src/result.py
@@ -1,7 +1,6 @@
-from utils import print_log
-from utils import GlobalSettings
-from state import State
-from save_system import load_history_list, load_history_data
+from src.utils import print_log, GlobalSettings
+from src.state import State
+from src.save_system import load_history_list, load_history_data
 
 
 class ShowHistory(State):
